backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=4.5.4
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5
email-validator>=2.0.0
httpx>=0.24.0 